chore: move FormChangeFunctions
